% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basv_prms.R
\name{basv_prms}
\alias{basv_prms}
\title{basv_prms}
\usage{
basv_prms(input, output, session)
}
\arguments{
\item{input}{shiny}

\item{output}{shiny}

\item{session}{shiny}
}
\value{
list of data and paramter
}
\description{
server backend to get a fieldbook
}
